# docs/assets/openapi.yaml
openapi: 3.1.0
info:
  title: DeskBot API
  version: 0.1.0
  summary: REST API for monitoring and controlling DeskBot.
  description: |
    HTTP API for monitoring and controlling a DeskBot instance.

    The API exposes robot state, configuration, perception, audio and
    conversation status, conversational history, speech and emotion commands,
    hardware calibration, user preferences, learning, and performance
    profiling endpoints.

    The WebSocket event stream is documented with the `x-websocket` extension
    because standard OpenAPI does not define WebSocket operations.
servers:
  - url: http://localhost:8000/api/v1
    description: Local DeskBot API
...
paths:
...
  /preferences:
    get:
      tags: [preferences]
      summary: List all preferences
      description: |
        Returns all learned user preferences, sorted by confidence.
      operationId: listPreferences
      responses:
        "200":
          description: List of preferences.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreferenceListResponse"
        "404":
          description: Preference tracker not available or not enabled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
              example:
                detail: Preference tracker not enabled
  /preferences/{key}:
    get:
      tags: [preferences]
      summary: Get a specific preference
      operationId: getPreference
      parameters:
        - $ref: "#/components/parameters/PreferenceKey"
      responses:
        "200":
          description: Preference value.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreferenceItem"
        "404":
          description: Preference not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
              example:
                detail: "Preference 'name' not found"

    delete:
      tags: [preferences]
      summary: Delete a preference
      description: |
        Removes a preference by key. Returns the deleted key.
      operationId: deletePreference
      parameters:
        - $ref: "#/components/parameters/PreferenceKey"
      responses:
        "200":
          description: Preference deleted.
          content:
            application/json:
              schema:
                type: object
                required: [status, key]
                properties:
                  status:
                    type: string
                    example: deleted
                  key:
                    type: string
                    example: name
        "404":
          description: Preference not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
              example:
                detail: "Preference 'name' not found"
  /learning/status:
    get:
      tags: [learning]
      summary: Learning service status
      description: |
        Returns the current learning service status including experience counts,
        model loss, training cycle information, and model version.
      operationId: learningStatus
      responses:
        "200":
          description: Learning service status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningStatusResponse"
        "404":
          description: Learning service not available.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
              example:
                detail: Learning service not available
  /learning/preferences:
    get:
      tags: [learning]
      summary: Learned preferences
      description: |
        Returns preferences learned by the learning service, including
        confidence, observation count, and average reward for each pattern.
      operationId: learningPreferences
      responses:
        "200":
          description: Learned preferences.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningPreferencesResponse"
        "404":
          description: Learning service not available.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
              example:
                detail: Learning service not available
  /learning/config:
    get:
      tags: [learning]
      summary: Learning configuration
      description: |
        Returns the current learning service configuration including schedule,
        resource limits, and checkpoint settings.
      operationId: learningConfig
      responses:
        "200":
          description: Learning configuration.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningConfigResponse"
        "404":
          description: Learning service not available.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
              example:
                detail: Learning service not available
  /learning/train:
    post:
      tags: [learning]
      summary: Force a training cycle
      description: |
        Forces an immediate training cycle regardless of the configured
        schedule. Returns whether training was triggered and the current
        training state.
      operationId: forceTrain
      responses:
        "200":
          description: Training cycle result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningTrainResponse"
              examples:
                triggered:
                  value:
                    triggered: true
                    training_cycles_completed: 3
                    is_training: true
                rejected:
                  value:
                    triggered: false
                    training_cycles_completed: 2
                    is_training: false
        "404":
          description: Learning service not available.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
              example:
                detail: Learning service not available
...
components:
  parameters:
...
    PreferenceKey:
      name: key
      in: path
      required: true
      description: Preference key.
      schema:
        type: string
      example: name
...
  schemas:
...
    PreferenceItem:
      type: object
      required: [key, value, confidence, source, updated_at]
      properties:
        key:
          type: string
          description: Preference category (e.g. "name", "humour").
          example: name
        value:
          type: string
          description: Learned preference value.
          example: Alice
        confidence:
          type: number
          format: float
          description: Confidence score in [0, 1].
          example: 0.9
        source:
          type: string
          description: How the preference was learned ("explicit" or "inferred").
          example: explicit
        updated_at:
          type: string
          format: date-time
          description: When the preference was last updated.
    PreferenceListResponse:
      type: object
      required: [preferences]
      properties:
        preferences:
          type: array
          items:
            $ref: "#/components/schemas/PreferenceItem"
    LearningStatusResponse:
      type: object
      required: [enabled, total_experiences, new_experiences_since_train, training_cycles_completed, is_training, model_version]
      properties:
        enabled:
          type: boolean
          example: true
        total_experiences:
          type: integer
          example: 42
        new_experiences_since_train:
          type: integer
          example: 5
        training_cycles_completed:
          type: integer
          example: 3
        current_model_loss:
          type: number
          format: float
          nullable: true
          description: Current model loss, or null if not yet trained.
        candidate_model_loss:
          type: number
          format: float
          nullable: true
          description: Candidate model loss, or null if no candidate.
        last_training_time:
          type: string
          format: date-time
          nullable: true
          description: ISO timestamp of the last training cycle, or null.
        last_training_duration_s:
          type: number
          format: float
          example: 2.5
        is_training:
          type: boolean
          example: false
        promotions:
          type: integer
          example: 2
        rollbacks:
          type: integer
          example: 0
        model_version:
          type: integer
          example: 2
    LearnedPreference:
      type: object
      required: [key, category, value, confidence, observation_count, avg_reward, source, last_observed]
      properties:
        key:
          type: string
          example: volume
        category:
          type: string
          example: audio
        value:
          type: string
          example: "0.7"
        confidence:
          type: number
          format: float
          example: 0.85
        observation_count:
          type: integer
          example: 12
        avg_reward:
          type: number
          format: float
          example: 0.92
        source:
          type: string
          example: inferred
        last_observed:
          type: string
          format: date-time
    LearningPreferencesResponse:
      type: object
      required: [preferences, total_patterns]
      properties:
        preferences:
          type: array
          items:
            $ref: "#/components/schemas/LearnedPreference"
        total_patterns:
          type: integer
          example: 7
    LearningConfigResponse:
      type: object
      required: [schedule, resource_limits, checkpoint_config]
      properties:
        schedule:
          type: object
          required: [min_new_experiences, train_interval_s, min_experiences_for_training]
          properties:
            min_new_experiences:
              type: integer
              example: 10
            train_interval_s:
              type: number
              format: float
              example: 300.0
            min_experiences_for_training:
              type: integer
              example: 5
        resource_limits:
          type: object
          required: [batch_size, training_epochs_per_cycle, max_cpu_fraction, max_model_params, eval_sample_size]
          properties:
            batch_size:
              type: integer
              example: 32
            training_epochs_per_cycle:
              type: integer
              example: 3
            max_cpu_fraction:
              type: number
              format: float
              example: 0.3
            max_model_params:
              type: integer
              example: 10000
            eval_sample_size:
              type: integer
              example: 100
        checkpoint_config:
          type: object
          required: [checkpoint_dir, keep_last_n, promote_threshold]
          properties:
            checkpoint_dir:
              type: string
              example: /home/user/.deskbot/checkpoints
            keep_last_n:
              type: integer
              example: 3
            promote_threshold:
              type: number
              format: float
              example: 0.01
    LearningTrainResponse:
      type: object
      required: [triggered, training_cycles_completed, is_training]
      properties:
        triggered:
          type: boolean
          example: true
        training_cycles_completed:
          type: integer
          example: 3
        is_training:
          type: boolean
          example: true
...
    HTTPError:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
...